+++ vars/vars.go
@@ -2,7 +2,7 @@ package vars
 
 var (
 	AppName    = "go-cli-starter"
-	AppVersion = "1.0.0"
+	AppVersion = "v1.1.0"
 	GoVersion  = "default"
 	BuildTime  = "default"
 	GitCommit  = "default"